requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
Django==4.0.5
django-cors-headers==3.13.0
django-material==1.10.0
django-material-admin==1.8.6
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
